loader: Add aarch64 support to common
On aarch64 systems the kernel must be the first loaded module. The kernel is
loaded via `LOAD_RAW`, and when a `LOAD_RAW` is performed and no files are
yet loaded we allocate an extra page for loading the file. This allows
the loader to (later) align the file contents in such a way that the dboot
entry point is on a page boundary, which is required for the self-relocation
code to work.