loader: Add aarch64 support to libefi
Aside from Makefile support, this change tweaks an EFI header to
ensure that int types work for aarch64.

Worth noting is that the time source used (time_event.c) is chosen
because it uses UEFI Boot Services. The implementation in time.c uses
UEFI Runtime Services - we can't use this as we may not be at the final
exception level, and calling Runtime Services from different exception
levels is unsupported.