loader: Add aarch64 arch support for EFI loader
Some of this change are direct imports from FreeBSD (cache.c, cache.h).

The existing FreeBSD linker script had a few flaws in our environment.

`nullconsole.c` and `spinconsole.c` are copies of the code in libi386,
and they're not at all architecture-sensitive.

The `start.S` file includes a handcrafter PE header, which allows
us to produce an EFI application via `objcopy`. Having a working PE
conversion in binutils would be ideal, but this will do for now.

`start.S` was also updated to use pc-relative addressing, ensuring that
all early code is position independent.

The `exec.c` file has support for using the loaded kernel (the whole
file), resolving symbols in that file and simply jumping to the loaded
dboot entry point. Loading of the file uses new new "extra page" load
method, allowing the dboot entrypoint to be page-aligned, which is
a prerequisite of the dboot self-relocation code.