loader: Add MMIO UART scaffolding and pl011 support
The MMIO UART framework abstracts UART implementations from the rest
of `loader(7)`, supporting simplified UART implementations (only
`pl011` is currently implemented) while implementing the expected
`loader(7)` behaviour only once.

This framework contains a set of flags to guide the logic of fetching
and setting UART configuration. There are three types of configuration:
1. Firmware-specified, pre-programmed settings
2. Hardware-specified, pre-programmed settings
3. Software-specified settings

Case 1 deals with settings specified by the FDT on a
`/chosen/stdout-path` node and the ACPI SPCR. These settings should
never be changed (but: see below for an escape-hatch).

Case 2 deals with a lack of settings specified by the FDT on a
`/chosen/stdout-path` node. The settings should be derived from what
the loader (u-boot) has programmed into the hardware.

Case 3 is for regular UARTs, which should be programmed in the usual
way.

In locked-settings cases (1 and 2), the user can set the `ignore-spcr`
environment variable to "true" to override the lock and program the
desired settings. This, however, is discouraged and one should prefer
updating the firmware with the desired settings.

This framework is intended to be driven by firmware-specific
configuration logic, added in subsequent diffs.