loader: Discover consoles from FDT
This change implements UART discovery from FDT. Only `pl011` is
supported at the moment. This library drives the MMIO UART library
for the implementation of the UART drivers.

Discovery is driven by looking through the /aliases node for entries
matching `serial0`-`serial25`. The serial index becomes the `tty` name,
with `serial0` becoming `ttya`, `serial1` becoming `ttyb` etc.

The code uses `/chosen/stdout-path` to determine which UART is the
system stdout. UART configuration is parsed from this `/chosen` node.
If configuration was present, then the UART is set as having
firmware-specified configuration. If no configuration was present, then
the UART is set as having hardware-specified configuration.

Clock frequency is required for setting UART configurations, so the
library implements clock frequency resolution for `fixed-clock` and
BCM2835 CPRMAN clocks (a.k.a Raspberry Pi 4). Frequency resolution
falls back to the `clock-frequency` property on the UART node, which
is common in simple FDT trees.

The BCM2835 CPRMAN implementation uses the BCM2835 MBOX facility, which
also requires dcache invalidation for a range of addresses, so this
diff adds that code.

This diff also includes the required register and ranges resolution
and dma-ranges resolution (both bus-to-cpu and cpu-to-bus), as needed
by the clock resolution code.

Finally, this diff hooks up the FDT UART discovery in the consoles
table.