loader: Discover consoles from the ACPI DSDT
Similar to the FDT UART discovery library, this diff adds UART discovery
from the ACPI DSDT.

The DSDT is traversed, looking for compatible UARTs. When a compatible
UART is found, the library walks the resources for the device,
extracting the MMIO regions, then walks the device-specific data,
looking for the properties UUID. If the properties UUID is found, the
library attempts to extract the `clock-frequency` for the device when
one is present.

When a device matches the SPCR, the device attributes are set per the
configuration in the SPCR and, if no `clock-frequency` was found, the
clock frequency is set to the value in the SPCR (if the SPCR is a new
enough version that it contains this field).

Once the system bus has been traversed, the library attempts to assign
device indices to the discovered devices. The most common method of
assigning indices is when the _HID value for all discovered devices is
the same, in which case the ACPI _UID can be used as a source of port
numbering. Failing that, the code attempts to derive nunbering from
the device node names when all name stems are the same.  If all
index-assignment logic fails the devices are assigned indexes in
discovery order.

Once indexes are assigned the library creates MMIO UART objects for
each discovered device in the usual way, making the UARTs visible as
consoles to loader.