armv8: Move supported boards to be platmods
Makefiles have been updated with the `platmod` machinery based on the
`sun4v` implementation and `sun4v/ontario` platform. The
`LINKED_PLATFORMS` code has been removed from the install rules so
that (unused) userland directories are not created. The approach taken
here, for the most part, has been to reapply the `rpi4` and `virt`
changes on top of the makefiles from `uts/sun4v/ontario` and to pull
the relevant build machinery into `uts/armv8` from `uts/sun4v`.

Since we already have a `sys/platform_modules.h` in `uts/aarch64`,
which is used by code (via unimplemented weak interfaces) in `genunix`,
I have elected to keep the `sys/platmod.h` in `armv8` as the place
to define the platform interfaces needed by the port, given these are
`armv8`-specific (and in most cases, `rpi4`-specific). The functions
are now declared as weak and checked using the usual pattern at
the callsites.

The platform directories remain unchanged, but a link has been
added to the name that would be defined by the first compatible
entry on the root node of the passed FDT. This allows us to experiment
with removing the "vanity naming" of the root node and make naming
be driven by firmware tables alone.

With the platmod code extracted (`virt/os/virt.c`, `rpi4/os/rpi4.c`)
the kernel is identical between the two platforms. Given this, move
the kernel to live in `uts/armv8`. The kernel is located via the
existing default path fallback mechanism in `inteboot` - no code
changes were needed here.

The platform modules have been updated to elide unimplemented
functions, implement necessary data structures (`platform_module_list`)
and expected functions that are not checked for presence
(`plat_tod_fault`).