devfdt: Add /dev/fdt exposing the flat device tree blob to userland
This is entirely for observability, and doesn't implement anything the tools
we have don't seem to need.  Most especially, you can't mmap(2) this device,
though it would be easy to add.

Longer term, it is possible that really we should extend bootfs to have a
notion of correct file permissions, and loader should create /system/boot/fdt
as a boot module.

Since the boot-time FDT contains theoretically security-relevant stuff (the
kaslr and rng seeds, if we used them), the device is 0400 but not otherwise
protected.  This matches openprom(4D) which had just about the same problem,
on SPARC, I think.