boot: Implement shared boot functionality
Similar to i86pc, implement shared boot functionality (console and
reset) so that this code can be shared between a boot stub (like
dboot) and the OS.

While this change includes framebuffer code, the framebuffer code
will not be used at present, as inetboot does not pass a framebuffer
to unix. The framebuffer will come when we move to loader(7). In order
to make this all hang together, where the text console was being
assumed, bcons_init now looks at the passed xboot_info the set
defaults based on what the bootloader passed.

There is no notion of selecting a specific serial output. In aarch64
systems the boot UART (the one selected via `/chosen/stdout-path` or
via the ACPi SPCR table) is passed to `unix` by the bootloader, so
that is always the UART used by the boot code. To change this serial
output device one can change the FDT used to boot the machine, much
like you would need to do in other operating systems.

Keyboard functionality is stubbed out since we don't yet know how
we'll be handling keyboards this early in boot.

Something to be aware of is that with this change `unix` no longer
relies on the bootloader having set up device and FDT mappings in
KPM-space. The FDT is reallocated into normal kernel memory and
the prom reinitialised while the UART is mapped into the device
arena once that is available.