boot: Move boot_archive to an IPD 34 compliant location
Out boot_archive location appears to reflect the strong SPARC influence
in the original port, and as such does not include the ISA directory
in the path.

So, to support treating the two modern illumos platforms as similarly
as possible, move the boot archive into a location that can be
constructed identically between the two platforms.