kmdb: don't restore all of the spsr after a single-step
When single stepping, we need to flip some processor state to enable stepping,
disable interrupts and debug traps.

Unfortunately, this was being done by saving and restoring the spsr across the
step, which clobbers the condition codes of the stepped instruction.  If
that's a `cmp` awful things will happen.

Many thanks to Andy Fiddaman for debugging this.

Instead, take the spsr from after the step, and clear the flags from _that_.
We could try to be more cautious and restore the pre-step state of SS,I,D but
I think that just makes things more complex.