kmdb: deferred breakpoint resolution can enter infinite loop
When using deferred breakpoints on rpi4, kmdb never recovers
after resolving the breakpoints following a module load. Despite
replacing the `brk #0x0` instruction at the start of
kmt_defbp_enter_debugger with the original before resuming, the
system behaves as if the original brk instruction is still there.

The loop looks like:

    mdb DEBUG: brkpt disarmed at fffffffffe4d54d0 kmt_defbp_enter_debugger
    mdb DEBUG: Resume requested, pc is fffffffffe4d54d0
    mdb DEBUG: returning to driver - cmd 2 (work required)
    mdb DEBUG: Back from resume, pc: fffffffffe4d54d0, trapno: 60
    mdb DEBUG: returning from resume

Note that we continue to get trap 60 and the PC does not advance.

Flushing both the data and instruction caches for the address of
the patched instruction, as well as data and instruction barriers,
resolves the problem.