dacf: Flesh out consplat for aarch64
Flesh out the platform console selection for armv8/aarch64 platforms.

This change merges in logic from the `i86pc` implementation, so
copyrights are updated accordingly.

Alongside those changes, we recognise `ttya`-`ttyz`, as there's no
architected notion of built-in consoles as one might find on `i86pc`.

The text (framebuffer) console is recognised, as are keyboards and mice,
none of which are in place yet, but will be put up for review shortly.

This is necessary so that we respect the console settings provided by
loader, rather than looking at the FDT.