armv8: Create device tree node for efifb
When a bootloader has passed a usable UEFI framebuffer to `unix` we're
going to want to attach a framebuffer console to that memory.

Create a trivial device node for attaching this driver under the root
of the device tree.

PCIe/PCI framebuffers are an open question that we can only answer
later.