boot: Relocate the boot console font
As with `i86pc`, the memory that the boot font was loaded into is a
temporary mapping that will be destroyed as the kernel comes up.

Import the `i86pc` code that relocates the boot font and hook it up
in the same place as we do on `i86pc`.