boot: Pass the SMBIOS3 pointer to unix
SMBIOS3, while mandatory in SBBR, is optional for the illumos kernel.
Pass the address of the SMBIOS3 table to `unix` from `dboot` when the
configuration table is present in the UEFI System Table.

This configuration table is passed to us by both U-Boot and traditional
UEFI firmwares, so we can expect to see this populated on all currently
supported systems.