dboot: Don't lose information on the kernel command-line
Based on an email discussion, dboot swallows unrecognised kernel
arguments, which makes it impossible for us to move boot verbosity
control to loader configuration rather than code changes.

The definitive testcase is:
```
ok boot -vm verbose
```

In this case dboot mangles the command-line such that the kernel sees:
```
-v verbose
```

This change stops processing the command-line when the first
unrecognised argument is encountered, allowing the code block that
handles copying out unrecognised arguments to kick in. This had been
identified by the compiler as dead code, shown by the need to switch
`bcopy` (unimplemented in dboot) to `memcpy`.

There are three test cases.

`boot -Vv -- -m verbose` shows how to use end-of-arguments to construct
this command-line. This works both before and after this change.

`boot -Vv -m verbose` shows the problem, where `-m` is stripped. This
fails without the change and passes after the change.

`boot -Vvm verbose` also shows the problem, where `m` is stripped. This
fails without the change and passes after the change.

A final test shows that dboot fails to strip out the only dboot argument
when that argument comes after an unrecognised option. This test is
`boot -vm verbose -V`, and results in `krtld` ignoring the
dboot-specific option.