unix: Integrate SMBIOS
Integrate a simplified version `os/smb_dev.c`, plucked from `i86pc`.
The edits to this file center around removing support for SMBIOS
versions prior to 3.0, not falling back to DOS-style memory scanning
for locating SMBIOS tables and ensuring that `CE_CONT` messages are
newline terminated.

The integration code in `os/startup.c` is a copy of that from `i86pc`,
with the conditionalisation for Xen removed.