ufsops: Treat kmem_ready as a weakish symbol
When booting from a UFS bootrd the `bufs_close` function references
the `kmem_ready` symbol (from `kmem.o` in `genunix`) to decide how
the inode cache should be managed.

Since this all happens very early in boot, while loading and linking
`genunix`, this symbol is not yet resolved. Use the same trick as
can be found in `krtld/bootrd.c` to check the address of the
`kmem_ready` global prior to attempting to access it.

With this change we can boot from an uncompressed UFS ramdisk:
```
ramdisk0 at root
ramdisk0 is /ramdisk
root on /ramdisk:a fstype ufs
```

We can also boot from ufs and ufs-uncompressed boot_archive files.