pciex/hotplug: need to avoid ddi_getlongprop, or to guard the warning we make it cause
We made the old type-less DDI property interfaces an error to use in the arm
port, so we have to avoid using it here.

Unfortunately that's made slightly more annoying in the non-PCIe case where
it's not 100% clear how we would interpret "slot-names".  On PCIe the bitmask
is always 0 and we don't have an issue.