aarch64: yield does not
Somewhat strangely, yield does not actually yield in all but a few
cases. Where it does work is in systems with CMT. Where it does not
work is... basically all machines you can buy that contain an Arm CPU.

What's worse is that the yield instruction lives in NOP-space, and is
therefore discarded by the CPU frontend on most modern Arm processors
(this is definitely the case on Neoverse-V2), so it's misleading at
best, and incorrect wherever it is used for things like spin-loop
backoff.

Advice from Arm is to use an instruction sync barrier (isb) to
introduce a short stall in processing. Terrible, but this is the
state of the Art until future CPUs ship with the wfet instruction.