armv8/vm: re-parameterize the HAT through the global `mmu.`
The x86 HAT indirects through the global `mmu.` to allow for runtime-variance
in the number of levels, page sizes, etc, and the scaling of the htable hash.

Restore all of this on ARM, as all of these can also vary there (though they
don't, right now).

There are some issues where I have replicated the x86 code but I'm unsure
about it:
  In cases where we check against the maximum level, and then decrement seem
  to be making the assumption that a page may always live at the
  (top level - 1), which is not true on ARM with larger virtual address
  spaces.

In doing this also:

  - bring in the smaller level_t and description from x86.
  - remove the duplicative MAX_* constants

The MAX_* constants exist for reasons I don't understand, it is apparent from
the code that the equivalent MMU_* constants _also_ represent the
architectural maximums (we use them to size arrays, and also on SPARC
explicitly say so).

Also document that `hw_page_array` is 0-terminated and so bigger than you'd
expect, and point to `page_szc` as evidence.