armv8: clean up static references to segkpm
We should refer to segkpm via the dynamic `kpm_vbase`, and/or calculate
addresses via the `hat_kpm_*` interfaces.  This will allow, in future,
kpm to transparently change address or size.

While here, also assert that kpm exists.  It is not optional on arm (as
it used to be optional, or at least missing, on 32bit platforms).